WIP/RESOVE MERGE-2 with Stanley Dec 2022 >
STILL PARENT CRASH when init new 7070 modem (see bottom here):

```C
    int n = sprintf(m_buffer, "AT+CIPSEND=0,%u,\"%s\",%u\r", len, udpIP.c_str(), udpPort);
```

SAME crash suffered parent Hard Merge!

- Preserve `kph` & `deviceTemp` as float (vs Stanley's integers)
- Skip sending `PID_GPS_HEADING` & `PID_ACC`  (vs Stanley's).
- Fix bug in external sensors from previous hard-merge.

---------
```
D (10873) APPUDP: notify: |UCFLV8Y2#EV=7,TS=9680,ID=UCFLV8Y2*95|
V (10879) APPUDP: notify x0...
Guru Meditation Error: Core  1 panic'ed (StoreProhibited). Exception was unhandled.

Core  1 register dump:
PC      : 0x40087d14  PS      : 0x00060c30  A0      : 0x801550b0  A1      : 0x3ffb2170
A2      : 0x00000000  A3      : 0x3f405b8a  A4      : 0x0000000d  A5      : 0x3f41e05c
A6      : 0x00000001  A7      : 0xff000000  A8      : 0x00000000  A9      : 0x00000000
A10     : 0x00000041  A11     : 0x00000000  A12     : 0x3ffb2434  A13     : 0x00000000
A14     : 0x3ffb6f7c  A15     : 0x00000008  SAR     : 0x00000004  EXCCAUSE: 0x0000001d
EXCVADDR: 0x00000000  LBEG    : 0x40088f25  LEND    : 0x40088f35  LCOUNT  : 0xfffffffe

Backtrace:0x40087d11:0x3ffb21700x401550ad:0x3ffb2190 0x4014b346:0x3ffb21c0 0x40146e9d:0x3ffb24d0 0x400e32be:0x3ffb2590 0x400db1d2:0x3ffb25c0 0x400de48e:0x3ffb26f0 0x400e0770:0x3ffb27c0 0x400e7336:0x3ffb2820

  #0  0x40087d11:0x3ffb2170 in memmove at /builds/idf/crosstool-NG/.build/xtensa-esp32-elf/src/newlib/newlib/libc/string/memmove.c:75
  #1  0x401550ad:0x3ffb2190 in __ssprint_r at /builds/idf/crosstool-NG/.build/xtensa-esp32-elf/src/newlib/newlib/libc/stdio/vfprintf.c:340
  #2  0x4014b346:0x3ffb21c0 in _svfprintf_r at /builds/idf/crosstool-NG/.build/xtensa-esp32-elf/src/newlib/newlib/libc/stdio/vfprintf.c:1773 (discriminator 1)
  #3  0x40146e9d:0x3ffb24d0 in sprintf at /builds/idf/crosstool-NG/.build/xtensa-esp32-elf/src/newlib/newlib/libc/stdio/sprintf.c:618
  #4  0x400e32be:0x3ffb2590 in CellUDP::send(char const*, unsigned int) at /home/ankostis/Work/JRCMATICS.git/libraries/FreematicsPlus/FreematicsNetwork.cpp:749
  #5  0x400db1d2:0x3ffb25c0 in TeleClientUDP::notify(unsigned char, char const*) at teleclient.cpp:265 (discriminator 10)
  #6  0x400de48e:0x3ffb26f0 in initialize() at telelogger.cpp:659
  #7  0x400e0770:0x3ffb27c0 in setup() at telelogger.cpp:1888
  #8  0x400e7336:0x3ffb2820 in loopTask(void*) at /home/ankostis/.platformio/packages/framework-arduinoespressif32/cores/esp32/main.cpp:42

ELF file SHA256: 0000000000000000

Rebooting...
```

+++ libraries/FreematicsPlus/FreematicsPlus.cpp
@@ -139,7 +139,7 @@ static void gps_decode_task(void* inst)
 }
 
 // get chip temperature sensor
-int readChipTemperature()
+float readChipTemperature()
 {
 #ifdef ARDUINO_ESP32C3_DEV
     static bool inited = false;
